refactor: findPath를 MoveStrategy 인터페이스로 분리
