+++ src/main/java/janggi/domain/movestrategy/GuardStrategy.java
@@ -1,8 +1,10 @@
 package janggi.domain.movestrategy;
 
-import static java.lang.Math.abs;
-
 import janggi.domain.board.Position;
+
+import java.util.List;
+
+import static java.lang.Math.abs;
 
 public class GuardStrategy implements MoveStrategy {
 
@@ -22,5 +24,10 @@ public class GuardStrategy implements MoveStrategy {
         }
 
         return abs(preX - nextX) + abs(preY - nextY) <= 1;
+    }
+
+    @Override
+    public List<Position> findPath(Position from, Position to) {
+        return List.of(to);
     }
 }
