+++ src/main/java/janggi/domain/movestrategy/MoveStrategy.java
@@ -2,7 +2,11 @@ package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
 
+import java.util.List;
+
 public interface MoveStrategy {
     boolean canMove(Position from, Position to);
+
+    List<Position> findPath(Position from, Position to);
 }
